Refactor specs for candidate reviewing and submitting an application.
