+++ spec/system/candidate_interface/candidate_entering_personal_details_spec.rb
@@ -1,6 +1,9 @@
 require 'rails_helper'
+require_relative 'helpers/candidate_helper'
 
 RSpec.feature 'Entering their personal details' do
+  include CandidateHelper
+
   scenario 'Candidate submits their personal details' do
     given_i_am_not_signed_in
     and_i_visit_the_personal_details_page
@@ -29,21 +32,6 @@ RSpec.feature 'Entering their personal details' do
 
     when_i_click_on_personal_details
     then_i_can_check_my_revised_answers
-  end
-
-  def given_i_am_not_signed_in; end
-
-  def and_i_visit_the_personal_details_page
-    visit candidate_interface_personal_details_edit_path
-  end
-
-  def then_i_should_see_the_homepage
-    expect(page).to have_current_path(candidate_interface_start_path)
-  end
-
-  def given_i_am_signed_in
-    candidate = create(:candidate)
-    login_as(candidate)
   end
 
   def and_i_visit_the_site
