+++ spec/system/candidate_interface/candidate_viewing_application_spec.rb
@@ -1,15 +1,13 @@
 require 'rails_helper'
+require_relative 'helpers/candidate_helper'
 
 RSpec.feature 'Viewing their new application' do
+  include CandidateHelper
+
   scenario 'Logged in candidate with no application choices' do
     given_i_am_signed_in
     when_i_visit_the_site
     then_i_should_see_that_i_have_made_no_choices
-  end
-
-  def given_i_am_signed_in
-    candidate = create(:candidate)
-    login_as(candidate)
   end
 
   def when_i_visit_the_site
